fix(spring): address review feedback on declarative AOP annotations
  - Add RecordingTxManager to H2 aspect tests to prove the named bean is
    actually selected (Result, Try, Validated variants)
  - Remove redundant .intValue() in Postgres aspect tests (Try, Validated)
  - Split duplicate public class OrderService into two code fences in the
    programmatic-vs-declarative guide snippet
  - Clarify Spring Boot @ComponentScan caveat for DmxTransactionalAspect
  - Fix timeout attribute description: "platform default" → TIMEOUT_DEFAULT (-1)
  - Add missing imports to aspect-setup.mdx guide snippet

+++ spring/src/test/java/dmx/fun/spring/TransactionalTryAspectPostgresTest.java
@@ -54,7 +54,7 @@ class TransactionalTryAspectPostgresTest {
     }
 
     private int countRows() {
-        return jdbc.queryForObject("SELECT COUNT(*) FROM events", Integer.class).intValue();
+        return jdbc.queryForObject("SELECT COUNT(*) FROM events", Integer.class);
     }
 
     @Test
